requirements.txt
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.2
pandas==2.0.3
scipy==1.11.1
dataclasses-json==0.5.14
pyyaml==6.0
pytest==7.4.0
jupyter==1.0.0
